Saving progress toward getting G4 simulation running within art.
MCTruth now just has a list of particles.  Element 0 of the particle vector is the beam particle.  Not sure if we want to keep it this way, we'll see how things go with Geant4 later.

Updated G4Gen producer module to take beam particle, create a MCTruth object, pass that to the G4Helper, and write out SSD hits and any new particles created by Geant to the event.