Factorize the unpacker class into two: one for artdaq fragments, another for SSD data.
Added new SSDUnpacker, so we can create art files for SSD data only.  Note, I made no attempt to extract and store useful the timestamps for the art events.  If that's important, we can do that later (eg, maybe based off the SSD raw data timestamps).

This works, and is ready to be merged with main.