Update/fix/add missing methods in rb::TrackSegment and rb::Track.
Updated Robert's MakeSingleTracks_module to record rb::tracksegments and rb::tracks to the art::event, in addition to rb::spacepoints.  A beam track is created based on the track segment in the two most upstream stations, and a single "secondary" track is created with the two other track segments. Note, however, that the momenta of the track segments is really just unit vectors, and the momenta and vertices of the tracks need to be updated.  Notes are in the code.