Oh, where to start...
Fixed a bunch of bugs related to the Align class and service.  The DetGeoMap class now gets a pointer to both a Geometry and an Align object (provided by their services).

Removed a bunch of unnecessary #includes in G4EMPH code (and fixed a typo in the name of a file).